+++ CHANGELOG.md
@@ -17,6 +17,7 @@
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateCoverOperateService。
   - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateParticipantOperateService。
+  - com.dwarfeng.familyhelper.life.stack.service.ActivityTemplateFileOperateService。
 
 - 新增实体及其维护服务，单元测试通过。
   - com.dwarfeng.familyhelper.life.stack.bean.entity.Activity。